Major changes
Bones now load to all MeshBoneLoaders in the scene, allowing multiple meshes to be parented to the bone structure. MeshBoneLoaders now have a bool useCollider, which determines if it spawns and uses colliders for the auto parenting (high vertex meshes will crash if they have to spawn all the colliders). Autoparenting can now work without using the vertex colliders, can use just the vertex. Holding x switches to the see through material which shows bone parenting, and letting go reverts it to it's original material.